get rid of utils & excess comments
